+++ app/assets/javascripts/getAvailability.js
